@@ -68,7 +68,7 @@ pytest-asyncio = {version = "^0.15.1", optional = true}
 pytest-cov = {version = "^2.12.1", optional = true}
 pytest-depends = {version = "^1.0.1", optional = true}
 pytest-lazy-fixture = {version = "^0.6.3", optional = true}
-python = "^3.8"
+python = "^3.10"
 python-multipart = "^0.0.5"
 rollbar = "^0.16.2"
 sentry-sdk = "^1.3.1"
