@@ -1,5 +1,4 @@
 from datetime import datetime
-from typing import Optional
 
 from pydantic import validator
 from sqlmodel import Field
@@ -20,16 +19,14 @@ from joj.horse.schemas.base import (
 
 
 class DomainInvitationEdit(BaseModel, metaclass=EditMetaclass):
-    code: Optional[LongStr] = Field(None, description="invitation code")
-    expire_at: Optional[UTCDatetime] = Field(
-        None, description="expire time of invitation"
-    )
-    role: Optional[str] = Field(None, description="domain role after invitation")
+    code: LongStr | None = Field(None, description="invitation code")
+    expire_at: UTCDatetime | None = Field(None, description="expire time of invitation")
+    role: str | None = Field(None, description="domain role after invitation")
 
 
 class DomainInvitationBase(URLORMSchema):
     code: LongStr = Field(index=True, nullable=False, description="invitation code")
-    expire_at: Optional[datetime] = Field(
+    expire_at: datetime | None = Field(
         None,
         sa_column=get_datetime_column(nullable=True),
         description="expire time of invitation",
@@ -43,7 +40,7 @@ class DomainInvitationBase(URLORMSchema):
 
 
 class DomainInvitationCreate(URLCreateMixin, DomainInvitationBase):
-    expire_at: Optional[UTCDatetime]
+    expire_at: UTCDatetime | None
 
     @validator("role", pre=True)
     def validate_role(cls, v: str) -> str:
